Get unarranged siblings
What _is_ a graph? A vector connected to other graphs.

Start with a graph key to create a vertex and set its point to be the
middle of w and h. Kick things of by arranging the graph with the first
vector.

Arranging a vector is, so far, getting it's sibling verteces that aren't
arranged yet. Once we have those we'll give them points and then draw
lines to them. For each vertex we'll then call arrange_vector on it
recursively.

I still need to calculate the points for each sibling vector, then
create the lines, then connect vectors that are being arranged back to
vectors already arranged:
  e.g. 1 -> 2 -> 3 -> 1
If we draw the graph0 {1 -> <graph1>} then we'll draw 1 connected to 2.
Then if we draw the grapn1 {2 -> <graph2>} we'll draw 2 connected to 3.
Then if we draw the graph2 {3 -> <graph3>} we won't draw anything.
_Unless_ we also draw the connection from 3 back to the vector 1 which
has already been arranged.
Eventually this will need to be pretty so that it doesn't look like 1, 2
and 3 are all in a straight line:

       1
      / \    instead of  1---2---3 (where 3->1 overlaps 2)
     2---3

I'm keeping a lot of the old code I had. I might need it but I'll delete
it as I replace it.