Make animation work: plumb in shapes, fix state
I just want to get this into a state where I can hit "animate" and
nothing breaks.

The width and height are passed as part of the render state. So far I
don't need them but I'll worry about that later. I'm trying to move
quickly so I can get a little work done in between other things.

The render function called by the animation framework relies on a list
of shapes in the animation state. This commit plumbs that in. It's
empty, but it's plumbed in.