Reduce requires entire state, not just shapes
I only have one accumulator to thread through the graph with the
reduce call. I'm including the shapes in the state along with other
necessary data and passing that in. Once it comes out we can ditch the
rest of the data (like unarranged keys) and just return the shapes in
the new animation state.

Once I start (_if_ I start) animating the graph to optimize over time
then I might need to keep some of the "arrange" state.