Only one reduce accumulator allowed
Fix the arrange_vectors param list so that we can use it as a reducer:
take the accumulator value as the first param, take the current object
as the second param.

Include graph in the state so that we can access it without having to
rely on the global variable, which might go away.