Only need to know arranged keys.
The shapes array will tell me where difrerent keys are, if I need to
know where they are later on. The arranged array is just to track which
keys don't need to be given x,y coordinates. Some keys will need to
connect to each other, but both will already have coordinates, so I'll
need to connect them "in place": i.e. the first (or "parent") key won't
dictate--directly--where the succeeding (or "child") key will be placed.

Example:
{1: [2],
 2: [3],
 3: [1]}
1 is placed
2 is placed relative to 1
3 is placed relative to 2
3 needs to connect to 1 but both are already placed, so for now we'll
just connect 1 and 3. This is not optimal because 3 will connect back to
1 in a straight line, intersecting 2.