Arrange children of already-arranged parents
If we arrange a parent in our breadth first pass and then swing back to
arrange its children, we should skip it because we've already arranged
it. That is, we shouldn't skip it altogether: we shouldn't re-arrange
it, but we should go on to arrange its children.

Example:

[{1: [2,3]},
 {2: [4]}]

Once we've arranged 2 and 3 as keys connected to 1 and get down to 2 as
a key in its own right, we shouldn't skip it because 4 hasn't been
arranged.

Also, record the origin vertex as already arranged so we don't arrange
it again. We're only adding a vertex to arranged list when we arrange it
within the arrange_shapes_ function. The first vertex is arranged before
that function is called.