Track connections between arranged verteces
Keep track of whether we've drawn the edge between two verteces that
were arranged separately: i.e. neither one drew the other. Example: 1 ->
2 -> 3 -> 1.
One draws two.
Two draws three.
Three needs to connect to one.
So, we need to keep track of which connections we have and if we don't
have 3->1 then we need to draw it and track it.