Show how verteces are constructed
The issue I'm seeing is because when I visit a vertex a second time,
after it's already arranged, I'm drawing it's connections using the
_second_ position and the angle of the _second_ parent.

In this case vertex 1 is arranging 2, 3, 4, 6 like so:
    2
    |
 3--1--6
    |
    4

... 1 has added 2, 3, 4, and 6 to the list of verteces to be arranged.
Next 2 is arranged and since 3 hasn't been arranged yet, it _also_ adds
three to the list to be arranged. (I think this should be changed).
However, it adds 3 based on it's own position and angle. Next the first
3 vertex ("3 from 1") adds 8, 9 and 10 to the list. Then later on the
second 3 vertex ("3 from 2") adds 8, 9 and 10 again, except this time
relative to its parent, which is 2.

To fix this I should either:
- not add a vertex to the list that's already in the list of vertexes to
  be arranged
- add the vertex to the list with the original parent's position and
  angle.

I think the first one makes more sense.