Comment out lots of debug
I had a lot of console.log statements to figure out drawing back
connections but I don't need it right now. I probably _should_ just
delete it, since it's all saved in previous commits, but I'm lazy and I
don't want to have to look it up if I need it in the near future.