WIP: does an edge overlap a vertex?
Go through each edge and check if that edge overlaps any vertex that
isn't one of the two that the edge connects:

+----+   +----+   +----+
|  B |___|__C_|___| D  |
|    |   |    |   |    |
+----+   +----+   +----+

If B connects to D then that B-D edge is overlapping C.

I'm going to check this by calculating the angles from the center of B
to the outer edges of C (B-to-C1, B-to-C2). Then I'm going to calculate
the angle from the center of B to the center of D. If the angle to the
center of D is between angles of B-to-C1 and B-to_C2 then the line from
B to D will pass through C. The line from B to D will go from the center
of B to the center of D.

Basically, I'm shooting some rays out tangential to C and seeing if I
catch D in those rays. If B and C are close then it will case a wide
beam but if B and C are far away it will cast a narrow beam. Said
another way, I'm seeing if D is in the shadow of C by casting a
point-source light from B and blocking it with C.

I'm currently in the process of calculating the top-most ray angle of C,
called "f" (because of a diagram I drew). Now I need to calculate the
angle of the other ray and then I can check the angle from B-center to
D-center. If the B-D angle is closer to one of the rays then I know
which direction to move D (or B, or C). Currently all of this is
assuming that B, C and D are in quadrant I and that B is bottom-left and
C is top-right; D can be anywhere. The gemoetry all changes depending on
where the vertices are in relation to each other.

"Make it work, make it right, make it fast."