WIP: calculating overlaps
Terms:

Length of vertex radius: r
1st vector in edge: ev1
2nd vector in edge: ev2
Vector that might be overlapped: ov
Side 1 of ov when ov faces ev1: ov1
Side 2 of ov when ov faces ev1: ov2

Assume that ev1 is the origin.

Take the absolute value of the difference between the x and y
cordinates of ev1 and ov. Let's refer to them as xd and yd.

Draw a line (L1) from the center of ev to the center of ov.

Draw a line (L2) perpendicular to the first line from the center of ov
towards the x axis.

Draw a line (L3) from the center of ov down to, and perpendicular to the x axis.

We can calculate the angle (A1) between the X axis and L1 by using the
Math.atan function and yd over xd.

Using geometry we know A1 will be the same as the angle (A2) between L3 and
L2.

Math.sin(A2) * r = distance from ov.x to ov1.x (positive) and ov2.x (negative)
Math.cos(A2) * r = distance from ov.y to ov1.y (negative) and ov2.x (positive)

Once we have the points ov1 and ov2 we can again use Math.atan to
calculate the angle from ev1 to the sides of ov (A3, A4).

Once we have the angles to the sides of ov, which form a "shadow", we
can use Math.atan again to get the angle of ev1 to ev2 (A5) and see if
A5 is between A3 and A4. That is, does ev2 lie in the shadow of ov with
respect to ev1.

There are a few gotchas:
If ev1 and ov or ev1 and ev2 are directly above, below or beside each
other (i.e. share x or y coordinates) then care has to be taken not to
divide by zero when using Math.atan.

If the ov1 crosses the x axis or ov2 crosses the y axis then care must
be taken to calculate the the angle from the other side of the axis
crossed.

If ov is in quadrant 1 in relation to ev1 and ov1 crosses over into
quadrant 4, then the "shadow" will be less than A3 and greater than A4.