Calculate moves based on overlaps
Instead of just calculated if an edge overlaps a vertex, I need to know
which vertex belonging to the edge needs to move, and in what direction.

Since the "shaadow rays" of a possibly overlapped vertex might straddle
zero degrees there's some logic to rotate both the shadow rays and the
edge angle counter-clockwise PI/2 radians so it's easier to deal with.

There's also some finagling to handle situations where the angle to move
the edge crosses zero degrees.

I have a bunch of console log statements because I forgot about the
return statement when the possiblly-overlapped vertex isn't even bounded
by the vertices attached to the edge. I was returning "false" and I
needed to return "undefined", as in "no move necessary"