WIP: fix the angle logic
The angle logic was all messed up.

I'm now considering "ov1" to be the angle to the side of the "maybe
overlapped" vertex that has the lowest angle; "ov2" is the angle to the
side that has the highest angle.

Things are looking good except the "5" vertex is moving and it shouldn't
be.