Wire up controls to play with inputs
I want to try goofy inputs to see what happens without having to edit
the source.

I copied a lot of the web page code from the js_doodle project, which
allows you to switch animations; this means the controls need to be
redrawn. In this case I don't need to redraw the controls.

There's a bug that causes moves to try and use angles that are greater
than 2*pi. I'm not sure where that's coming from but it appears when I
use "child spread" angle of ~0.1.