Use "link" log events to form a graph
When the MUD starts up each handler process has in its properties the
list of sibling IDs, but not their PIDs. We don't know what the PIDs
will be until each handler process is started. The link events are doing
an Erlang link() call to each sibling process by looking up the PID in
the ID->PID table. I'm using the link log events to figure out all
vertices (PIDs) and edges (PID linking to PID) in the graph.

The "link" log events have a "source" property, which is the linking
PID, and a "target" property which is the PID being linked to.