Convenience functions for "link" log events
I'm distinguishing between "populate", "link" and all other in-game
events. The "populate" event is when we're swapping all the IDs in the
handlers properties for the PID of the process for the object the
property was pointing to. The "link" event is for each one of those
ID->PID swaps we do. The rest of the log messages are for in-game event
steps: e.g. an event traversing the graph to see if a door can be
opened: each step will be an object in the graph handling that event.
Events can be "attempts" at an action, a "success" in completing the
action, a "fail" or various other types. I'm ignoring the "link" and
"populte" events to focus on the more complicated in-game events. The
logic is really hard to follow when everything is happening concurrently
but being logged to the same file. By drawing this out in 2D it will
hopefully be clearer.