Collapse "key" & "ID" concepts for clarity
Because I'm drawing a graph based on map keys I was referring to each
element that needed to be turned into a vertex as a key. However, when
creating shapes I was giving each shape an identifier which was the map
key.

When I got to creating a map of shapes to their coordinate points I was
confusing "key" with "id". I figure it's easier just to collapse these
down to one concept. Maps have keys, so calling them keys is accurate,
but the values I'm using for keys are IDs: once we're out of the context
of the original map it makes sense to refer to the underlying meaning of
the keys and not hold on to the fact that the IDs were used as keys in
the map.