PIDs must be stripped, coordinates exploaded
When the map to create the graph is first built in the graph.js:init()
function the link events are formatted with "stripped" pids. The graph
keys and values are then used to set the IDs of the shapes. Once we've
spread the shapes then we create an idPoints map of IDs to the X,Y
coordinates of the shape with that ID. Since the IDs from from links
with "stripped" PIDs we need to strip any PIDs that we use to look up
shape coordinates.