Draw event objects, fix rectangle, ...
draw_next_event() expects to send a drawable event (id, coordinates,
dimensions) to the draw_event() function.

We don't capture the event details in the draw_event function in a named
variable so we don't need to reference the values in the context of an
object.

Use the ID of the event since we no longer use the map key values of the
original graph as the shape IDs. (I did a search and replace s/key/id/
but I think I missed draw.js).