Vertex shapes don't actually move
When vertex shapes are "moved" I actually create a new shape at the new
location. Previously I was assuming that the animation was finished at
the end of animate_graph() and then get the coordinates of the original
shapes. However, the original shapes never move. Instead, I'm updating
the idPoints map every time a vertex shape is created with an ID at a
new location.