Show pending vertex moves before re-drawing
I want to see what the pending moves are for each vertex so I can try
and understand what the "spread" algorithm is doing. Obviously I can see
where each vertex moves to, but I want to see both where it is and where
it will be simultaneously.

This change draws a line from each vertex to its new point(s).

I'm not sure whether opposing moves cancel out or are applied in
sequence. However, I can see a lot of vertices that have two opposing
movements, or even three or four movements. That might be an issue.